tests: fix UBSan left-shift of negative literal in test_bsp_compiler_gtest
-1024 << 16 is UB for signed int; cast through uint32_t first.

+++ src/tests/unit/test_bsp_compiler_gtest.cpp
@@ -841,7 +841,7 @@ TEST_F(BspCompilerTest, OffsetOriginRoom)
     //   v0 +----+ v1
     //   Offset: x=-1024, y=-1024
 
-    int offset = -1024 << 16;
+    int offset = (int32_t)((uint32_t)-1024 << 16);
     std::vector<ZDBSPInputVertex> vertices = {
         {offset, offset},
         {offset + (256 << 16), offset},
